+++ misago/categories/permissions.py
@@ -6,7 +6,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from misago.acl import algebra
 from misago.acl.decorators import return_boolean
-from misago.core.forms import YesNoSwitch
+from misago.admin.forms import YesNoSwitch
 from misago.users.models import AnonymousUser
 
 from .models import Category, CategoryRole, RoleCategoryACL
