+++ misago/conf/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.utils.translation import ugettext as _
 from django.utils.translation import ungettext
 
-from misago.core.forms import YesNoSwitch
+from misago.admin.forms import YesNoSwitch
 
 
 __ALL__ = ['ChangeSettingsForm']
