+++ misago/threads/permissions/bestanswers.py
@@ -5,9 +5,9 @@ from django.utils.translation import ugettext_lazy as _, ungettext
 
 from misago.acl import algebra
 from misago.acl.decorators import return_boolean
+from misago.admin.forms import YesNoSwitch
 from misago.categories.models import Category, CategoryRole
 from misago.categories.permissions import get_categories_roles
-from misago.core.forms import YesNoSwitch
 from misago.threads.models import Post, Thread
