+++ misago/threads/permissions/polls.py
@@ -7,7 +7,7 @@ from django.utils.translation import ungettext
 from misago.acl import algebra
 from misago.acl.decorators import return_boolean
 from misago.acl.models import Role
-from misago.core.forms import YesNoSwitch
+from misago.admin.forms import YesNoSwitch
 from misago.threads.models import Poll, Thread
