+++ misago/threads/permissions/privatethreads.py
@@ -6,9 +6,9 @@ from django.utils.translation import ugettext_lazy as _
 from misago.acl import algebra
 from misago.acl.decorators import return_boolean
 from misago.acl.models import Role
+from misago.admin.forms import YesNoSwitch
 from misago.categories import PRIVATE_THREADS_ROOT_NAME
 from misago.categories.models import Category
-from misago.core.forms import YesNoSwitch
 from misago.threads.models import Thread
